# Conflicts:
#	routes/recipes.js

+++ routes/recipes.js
@@ -21,13 +21,13 @@ router.post('/create', auth, async (req, res) => {
     const { error } = validate(req.body);
     if (error) return res.status(400).send(error.details[0].message);
 
-    recipe = new Recipe(_.pick(req.body, ['name', 'ingredients', 'instructions', 'imageUrl', 'categoryId']));
+    recipe = new Recipe(_.pick(req.body, ['name', 'ingredients', 'instructions', 'image', 'categories']));
 
     // Author is the user that is creating the recipe.
     recipe.author = req.user._id;
 
     await recipe.save();
-    res.status(200).send(_.pick(recipe, ['_id', 'name', 'author', 'ingredients', 'instructions', 'imageUrl', 'categoryId']));
+    res.status(200).send(_.pick(recipe, ['_id', 'name', 'author', 'ingredients', 'instructions', 'image', 'categories']));
 })
 
 // Update recipe
@@ -40,21 +40,21 @@ router.put('/', auth, async (req, res) => {
     if (!recipe) return res.status(400).send('Recipe does not exist.');
 
     // If this isn't this user's recipe, return denied.
-    if (recipe.author != req.user._id) return res.status(403).send('No access to this resource.');
+    if ((recipe.author != req.user._id) || (!req.user.isAdmin)) return res.status(403).send('No access to this resource.');
 
-    let { name, ingredients, instructions, imageUrl, categoryId } = _.pick(req.body, ['name', 'ingredients', 'instructions', 'imageUrl', 'categoryId']);
+    let { name, ingredients, instructions, image, categories } = _.pick(req.body, ['name', 'ingredients', 'instructions', 'image', 'categories']);
 
     await Recipe.updateOne({ _id: recipeId }, {
         name: name,
         ingredients: ingredients,
         instructions: instructions,
-        imageUrl: imageUrl,
-        categoryId: categoryId
+        image: image,
+        categories: categories
     }, { omitUndefined: true });
 
     recipe = await Recipe.findOne({ _id: recipeId });
 
-    res.status(200).send(_.pick(recipe, ['_id', 'name', 'ingredients', 'instructions', 'imageUrl', 'categoryId']));
+    res.status(200).send(_.pick(recipe, ['_id', 'name', 'ingredients', 'instructions', 'image', 'categories']));
 });
 
 // Delete a recipe
@@ -64,10 +64,37 @@ router.delete('/:id', auth, async (req, res) => {
     if (!recipe) return res.status(400).send('Recipe does not exist.');
 
     // If this isn't this user's recipe, return denied.
-    if (recipe.author != req.user._id) return res.status(403).send('No access to this resource.');
+    if ((recipe.author != req.user._id) || (!req.user.isAdmin)) return res.status(403).send('No access to this resource.');
 
     await Recipe.deleteOne({ _id: req.params.id });
     res.send(recipe);
+});
+
+// Special Query
+router.post('/byFilter', async (req, res) => {
+    let { name, category, ingredient } = _.pick(req.body, ['name', 'category', 'ingredient']);
+
+    let query = [];
+
+    if (name && (name != "")) query.push({"name": { $regex: name }});
+    if (category && (category != "")) query.push({"categories": category});
+    if (ingredient && (ingredient != "")) query.push({"ingredients.name": { $regex: ingredient }});
+
+    let recipes = await Recipe.find({ $and: query });
+
+    res.send(recipes);
+});
+
+// Group by categories
+router.get('/categories', async (req, res) => {
+    let recipes = await Recipe.aggregate([
+        { $group: {
+            _id: "$categories",
+            data: { $push: { name: '$name', author: '$author', imageUrl: '$imageUrl', ingredients: '$ingredients', instructions: '$instructions', categories: '$categories'}},
+        } }
+    ]);
+
+    res.send(recipes);
 });
 
 module.exports = router;
